requirements.txt
openai
feedparser
gspread
oauth2client
python-dotenv
